+++ glean-core/ios/sdk_generator.sh
@@ -25,7 +25,7 @@
 
 set -e
 
-GLEAN_PARSER_VERSION=1.27.0
+GLEAN_PARSER_VERSION=1.28.1
 
 # CMDNAME is used in the usage text below.
 # shellcheck disable=SC2034