+++ samples/android/app/src/main/java/org/mozilla/samples/glean/MainActivity.kt
@@ -44,7 +44,6 @@ open class MainActivity : AppCompatActivity() {
             )
         }
 
-        uploadSwitch.setChecked(Glean.getUploadEnabled())
         uploadSwitch.setOnCheckedChangeListener { _, isChecked ->
             if (isChecked) {
                 gleanEnabledText.setText("Glean is enabled")